fix: extract knowledge data to separate file (max-lines lint), fix search state order
