+++ common/changes/@grackle-ai/cli/nick-pape-884-demo-mock-data_2026-03-26-03-57.json
@@ -1,7 +1,7 @@
 {
   "changes": [
     {
-      "comment": "placeholder",
+      "comment": "Enhanced demo mock data with realistic session events and knowledge graph",
       "type": "none",
       "packageName": "@grackle-ai/cli"
     }
